Use 16-bit equip pad masks

+++ src/menu_equip.cpp
@@ -323,8 +323,8 @@ bool CMenuPcs::EquipOpen0()
 int CMenuPcs::EquipCtrlCur()
 {
 	bool blocked = false;
-	unsigned int press;
-	unsigned int hold;
+	u16 press;
+	u16 hold;
 	CCaravanWork* caravanWork = reinterpret_cast<CCaravanWork*>(Game.m_scriptFoodBase[0]);
 
 	if ((Pad._452_4_ != 0) || (Pad._448_4_ != -1)) {
